+++ src/main/java/com/programmers/post/ui/dto/PostDto.java
@@ -1,4 +1,4 @@
-package com.programmers.post.dto;
+package com.programmers.post.ui.dto;
 
 public record PostDto(String title, String content, Long userId) {
 }
