+++ src/main/java/com/programmers/post/ui/dto/PostUpdateDto.java
@@ -1,4 +1,4 @@
-package com.programmers.post.dto;
+package com.programmers.post.ui.dto;
 
 public record PostUpdateDto(Long postId, String title, String content) {
 }
