+++ src/main/java/com/programmers/user/ui/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.programmers.user.dto;
+package com.programmers.user.ui.dto;
 
 public record UserDto(String name, int age, String hobby) {
 }
